Peer review and bug fix
